Improve update function, add Team Fortress Classic, Team Fortress 2 Classic, GTA: FiveM, GTA: San Andreas Multiplayer
