+++ sources/Mods/Mods.php
@@ -111,7 +111,8 @@ class _Mods extends \IPS\Node\Model
             'discord' => 'Discord',
             'gmod' => "Garry's Mod",
             'grav' => 'GRAV Online',
-            'gta5m' => 'GTA Five M',
+            'gta5m' => 'GTA: FiveM',
+            'samp' => 'GTA: San Andreas Multiplayer',
             'gtan' => 'Grand Theft Auto Network',
             'hl2dm' => 'Half Life 2: Deathmatch',
             'hurtworld' => 'Hurtworld',
@@ -134,7 +135,6 @@ class _Mods extends \IPS\Node\Model
             'quakelive' => 'Quake Live',
             'redorchestra2' => 'Red Orchestra 2',
             'rust' => 'Rust',
-            'samp' => 'San Andreas Multiplayer',
             'sevendaystodie' => '7 Days to Die',
             'ship' => 'The Ship',
             'squad' => 'Squad',
@@ -142,7 +142,9 @@ class _Mods extends \IPS\Node\Model
             'teamspeak3' => "Teamspeak 3",
             'teeworlds' => 'Teeworlds Server',
             'terraria' => 'Terraria',
+            'tfclasic' => 'Team Fortress Classic',
             'tf2' => 'Team Fortress 2',
+            'tf2clasic' => 'Team Fortress 2 Classic',
             'tibia' => 'Tibia',
             'tshock' => 'Tshock',
             'unreal2' => 'Unreal 2',
